Splitting the dm residuals into two separate tasks, one for individual exposures and the other combining all the data. These will be put into a single pipeline.
